Fix memory leak, null bool vector, refcount discipline, and type consistency
- Move validity check before allocation in eval_dict_scalar_binop to
  prevent leak on error path (high)
- Add missing QO_BOOL_VEC and QO_TIMESTAMP_VEC cases in eval_builtin_null,
  fixing the pre-existing null_bool_vector test failure (high)
- Add qo_retain in make_function_value and release in caller to fix
  reference-stealing violation (high)
- Change long to int64_t in eval_take/eval_drop for portability (medium)
- Add missing (long) casts for int64_t in value_print.c (medium)
- Remove unused lexer.h include from internal.c (low)
- Remove spurious braces in alloc_data_vec (low)

+++ src/value_print.c
@@ -159,7 +159,7 @@ static void qo_print_internal(Qo q, int depth) {
             if (lv == QO_LONG_NULL) qo_printf("0N");
             else if (lv == QO_LONG_INF) qo_printf("0W");
             else if (lv == QO_LONG_NEGINF) qo_printf("-0W");
-            else qo_printf("%ld", lv);
+            else qo_printf("%ld", (long)lv);
             break;
         }
         case QO_TIMESTAMP: {
@@ -284,7 +284,7 @@ static void qo_print_internal(Qo q, int depth) {
                 if (v == QO_LONG_NULL) qo_printf("0N");
                 else if (v == QO_LONG_INF) qo_printf("0W");
                 else if (v == QO_LONG_NEGINF) qo_printf("-0W");
-                else qo_printf("%ld", v);
+                else qo_printf("%ld", (long)v);
             }
             break;
         }
